fix(runtime): resolve duplicate declarations and type conversion errors
- Remove duplicate FieldInfo and ExceptionInfo type declarations
- Fix uintptr vs int64 type mismatches in memory allocation
- Remove duplicate isPowerOfTwo and alignUp function definitions
- Clean up unused variables in layout.go
- Ensure consistency between layout and region memory modules

+++ internal/runtime/block_manager.go
@@ -125,7 +125,7 @@ func (bm *BlockManager) AllocateBlock(region *Region, size RegionSize, alignment
 
 	// Align total size
 	if alignment > 0 {
-		totalSize = RegionSize(alignUp(uintptr(totalSize), uintptr(alignment)))
+		totalSize = RegionSize(alignUp(int64(totalSize), int64(alignment)))
 	}
 
 	// Check if this is a large object allocation
